# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API TP1_Boisvert_Olivier",
        "version": "1.0"
    },
    "paths": {
        "/api/equipment": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Liste de tous les équipements",
                "operationId": "771e58f407e40ca413f92c85f944e760",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/equipment/{id}": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Afficher un équipement",
                "operationId": "546d2e88aec2a0d7af708399e7e20a1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Equipment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Équipement non trouvé"
                    }
                }
            }
        },
        "/api/equipment/{id}/popularity": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Afficher la popularité d'un équipement",
                "operationId": "f073a07513b9a2e31ca3067c2dccd504",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Equipment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Équipement non trouvé"
                    }
                }
            }
        },
        "/api/equipment/{id}/avgRental": {
            "get": {
                "tags": [
                    "Equipment"
                ],
                "summary": "Afficher le prix moyen de location d'un équipement",
                "operationId": "ebc8a865c946dcf9caae82d71625291e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Equipment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "minDate",
                        "in": "query",
                        "description": "Date de début",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "maxDate",
                        "in": "query",
                        "description": "Date de fin",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Équipement non trouvé"
                    }
                }
            }
        },
        "/api/reviews/{id}": {
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Supprimer un avis",
                "operationId": "89986343b5dbd2b909c2764aeb7e4096",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Review ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Avis non trouvé"
                    }
                }
            }
        },
        "/api/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Créer un utilisateur",
                "operationId": "1a63883edb751cb640a2f6f516236f49",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Jane",
                                        "maxLength": 50
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe",
                                        "maxLength": 50
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "janedoe@gmail.com",
                                        "maxLength": 50
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "1231231234",
                                        "maxLength": 12
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Modifier un utilisateur",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Jane",
                                        "maxLength": 50
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe",
                                        "maxLength": 50
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "janedoe@gmail.com",
                                        "maxLength": 50
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "1231231234",
                                        "maxLength": 12
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur modifié"
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Equipment",
            "description": "Equipment"
        },
        {
            "name": "Reviews",
            "description": "Reviews"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}